requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.3.0
py2app==0.7.3
python-dateutil==1.5
python-openid==2.2.5
requests==2.8.1
static3==0.6.1
wheel==0.24.0